requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pandas
numpy
scikit-learn
joblib
pydantic
flask
